+++ app/widgets/PublishBrief/publishbrief.js
@@ -1,4 +1,5 @@
 var PublishBrief = {
+    timeout: 0,
     togglePrivacy: function() {
         var checked = document.querySelector('#publishbrief form #open');
 
@@ -22,14 +23,26 @@ var PublishBrief = {
         var embed = document.querySelector('#embed');
         embed.value = url.value;
         embed.onchange();
+        PublishBrief.saveDraft();
     },
     clearEmbed: function() {
         document.querySelector('input[name=embed]').value = '';
         PublishBrief_ajaxClearEmbed();
+        PublishBrief.saveDraft();
     },
     checkEmbed: function() {
         var embed = document.querySelector('input[name=embed]');
         embed.onchange();
+
+        document.querySelector('form[name=brief]').onkeyup = function() {
+            if(PublishBrief.timeout) clearTimeout(PublishBrief.timeout);
+            PublishBrief.timeout = setTimeout(function () {
+                PublishBrief.saveDraft();
+            }, 1000);
+        };
+    },
+    saveDraft: function() {
+        PublishBrief_ajaxSaveDraft(MovimUtils.formToJson('brief'));
     },
     enableSend: function() {
         MovimUtils.removeClass('#button_send', 'disabled');