+++ CHANGELOG.md
@@ -18,6 +18,7 @@ v0.11.1 (trunk)
  * New UI for adding a contact through a Gateway (by singpolyma)
  * Allow users to clear their information on the instance and leave it properly
  * Add NSFW filter configuration
+ * Save Draft of publications in Publish and PublishBrief
 
 v0.11
 ---------------------------